+++ crates/metaltile-std/src/mlx/fp_quantized_nax.rs
@@ -329,7 +329,7 @@ use metaltile::test_kernel;
                 DType::U32,
             ))
             .expect(TestBuffer::from_vec("out", pack_f32(&expected), dt))
-            .grid_3d(n / 32, m / 32, 1, [128, 1, 1])
+            .grid_3d(n as u32 / 32, m as u32 / 32, 1, [128, 1, 1])
     }
 
     #[test_kernel(name = "mlx/fp_qmm_nax/f32_multi_k", dtypes = [f32], tol = 0.001)]
@@ -352,6 +352,6 @@ use metaltile::test_kernel;
                 DType::U32,
             ))
             .expect(TestBuffer::from_vec("out", pack_f32(&expected), dt))
-            .grid_3d(n / 32, m / 32, 1, [128, 1, 1])
+            .grid_3d(n as u32 / 32, m as u32 / 32, 1, [128, 1, 1])
     }
 }
